requirements.txt
numpy
minisom
Pillow
joblib
seaborn
matplotlib
datasets
streamlit
plotly